Fix the validation of closing brace for angle-bracket
If the angle-bracket component has block params the validation of the
closing brace didn't work previously.
Validation of the position of the `as |...|` block is pretty complicated
and it wasn't performed at all previously. I leave it out of the scope
of this change and just make the rule account for that extra line
required for the block params to properly validate the position of the
closing brace.